Added a "subtype" conditional test
I added a "subtype" conditional test to the <if> element. It's basically copy-and-paste from the "type" test for the <if> element. I added this, because the default CSL source types (defined in Appendix III of the specification) were not granular enough for my application focused on legal users. I don't want to break compatibility with style creators in general, so I don't want to create my own source types. Subtypes allow me to have very granular conditional tests without breaking compatibility with other styles.