Preliminary work on LogicValidator, InterfaceBuilder
I added two files to the repository. Interface.js contains two modules: SetLogic and InterfaceBuilder. The SetLogic module examines all possible paths through a CSL style to reach a <text variable> leaf-node, and compiles them into a list. The Interface Builder takes this compiled list, and transforms it into a User-Interface. The purpose of the user-interface is to tailor the visible fields to the source type of the bibliographic source.